Add cljdoc config/adjustments
Was previously failing on cljdoc because cljdoc assumed it should do
ClojureScript analysis because it a saw .cljc files.
Told cljdoc only Clojure should be analyzed in doc/cljdoc.edn.

I noticed codox config (which cljdoc does not look at) includes some
specific namespaces to document but existing code has some :no-doc
metadata already.

So made bold assumptions:

Always document codox specified namespaces:
- datalevin.client
- datalevin.core
- datalevin.interpret
- datalevin.search-utils

Else if namespace contains no :no-doc anywhere assume it should not be
documented and :no-doc added to namespace:
- datalevin.binding.graal
- datalevin.constants
- datalevin.datafy
- datalevin.lmdb
- datalevin.protocol
- datalevin.remote
- datalevin.sparelist
- pod.huahaiy.datalevin

My assumption was that if there is some var(s) with a :no-doc in a
namespace without a :no-doc, the namespace wants to be documented.
This added the following to API docs:
- datalevin.bits
- datelevin.main
- datalevin.search
- datalevin.server

Also added all .md docs I found to cljdoc list of articles.
This adds:
- bench/README.md
- native/README.md

The above could be entirely wrong, am happy to refine with you.