Create dynamic var for data readers
This supports the use case of users who have values with tagged literals
stored in their databases. Just as with clojure.core/*data-readers* and
clojure.core/default-data-readers, there is also a static var containing the
default readers to fall back to in contexts where the dynamic value has been
rebound, which is used to initialize the dynamic var.