Fix for 288
This is the start of a fix for 288. However, it seems there are two
levels of caching. The query results are cached and parsing the query
is also cached.

I've fixed the first level of caching (the result caching) by adding a
fn-object to the key. However, the query caching will be more
challenging to fix as we need to parse the query to resolve the
functions, and if we're parsing the query then there's not much point
caching it (why my test still fails).

One solution would be remove the query caching completely:

```
(defn- parsed-q
  [q]
  (or (.get ^LRUCache *query-cache* q)
      (let [res (dp/parse-query q)]
        (.put ^LRUCache *query-cache* q res)
        res)))
```

Would become:

```
(defn- parsed-q
  [q]
  (dp/parse-query q))
```

However, thinking about the problem some more. This bug only really
comes up when working at the repl and or in contexts where a function
gets rebound.

Depending on how critical the current elegant query caching is. It
might be simpler to just document invalidating the cache manually when
changing a function used in a query.

Thoughts?